backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.2.2
python-multipart==0.0.22
pydantic==2.12.5
groq==1.1.1
PyMuPDF==1.27.2
faiss-cpu==1.13.2
sentence-transformers==5.3.0
numpy==2.4.3
torch==2.10.0
